// backups/teste.1/types.ts
export type Language = 'pt' | 'en' | 'es' | 'it' | 'de';

export interface Airport {
  code: string;
  name: string;
  location: string;
  isAzores: boolean;
}

export interface Flight {
  id: string;
  airline: string;
  flightNumber: string;
  origin: string;
  destination: string;
  departureTime: string; // HH:mm
  arrivalTime: string;   // HH:mm
  price: number;
  status: 'A Horas' | 'Atrasado' | 'Embarque' | 'Cancelado';
  stops: number;         // 0 for direct
  duration: string;      // e.g., "2h 30m"
  layover?: string;      // e.g., "Escala em LIS (1h 20m)"
}